[Patchset 9] trace_processor: insert ftrace events into trace storage
This CL introduces the trace inserter class which correltates sched
switch events to compute durations using data in TraceStorage.
TraceParser utilises the inserter as an indirect way of inserting into
the trace storage.

Bug: 80416541

+++ src/trace_processor/trace_parser.cc
@@ -139,7 +139,7 @@ void TraceParser::ParseSchedSwitch(uint32_t cpu,
   uint32_t prev_pid = 0;
   uint32_t prev_state = 0;
   const char* prev_comm = nullptr;
-  size_t prev_comm_len = 0;
+  uint64_t prev_comm_len = 0;
   uint32_t next_pid = 0;
   for (auto fld = decoder.ReadField(); fld.id != 0; fld = decoder.ReadField()) {
     switch (fld.id) {
